extraer componente FiltroPeriodo con refs y lógica de date picker
