+++ frontend/src/views/public/administrador/administrador.jsx
@@ -1,23 +1,17 @@
-import { useRef } from 'react'
 import useEstadisticasAdmin from '@/hooks/use-estadisticas-admin.js'
 import StatCard from './stat-card.jsx'
 import SeccionBarras from './seccion-barras.jsx'
 import TopSelecciones from './top-selecciones.jsx'
 import RankingUsuarios from './ranking-usuarios.jsx'
 import AdministradorSkeleton from './administrador-skeleton.jsx'
+import FiltroPeriodo from './filtro-periodo.jsx'
+import PropuestasPorEstado from './propuestas-por-estado.jsx'
 import styles from './administrador.module.css'
 
 const TARJETAS_GLOBALES = (stats) => [
   { icono: 'bi-people-fill', numero: stats.totalUsuarios, label: 'Usuarios registrados' },
   { icono: 'bi-collection-fill', numero: stats.totalFiguritasPublicadas, label: 'Figuritas publicadas' },
   { icono: 'bi-stopwatch-fill', numero: stats.totalSubastasActivas, label: 'Subastas activas' },
-]
-
-const PROPUESTAS = [
-  { label: 'Pendientes', key: 'pendientes', color: '#d49a2c' },
-  { label: 'Aceptadas', key: 'aceptadas', color: '#198754' },
-  { label: 'Rechazadas', key: 'rechazadas', color: '#dc3545' },
-  { label: 'Canceladas', key: 'canceladas', color: '#6c757d' },
 ]
 
 const MODALIDADES = [
@@ -27,9 +21,6 @@ const MODALIDADES = [
 ]
 
 const Administrador = () => {
-  const desdeRef = useRef(null)
-  const hastaRef = useRef(null)
-
   const {
     stats, cargando, recargando, error,
     desde, setDesde, hasta, setHasta,
@@ -46,59 +37,15 @@ const Administrador = () => {
           </div>
           <p className={styles.heroSubtitulo}>Estadísticas de la plataforma Figus Mundial 2026</p>
 
-          <div className={styles.heroPeriodoFiltro}>
-            <div className={styles.heroPeriodoCampo} onClick={() => desdeRef.current?.showPicker()}>
-              <i
-                className="bi bi-calendar3"
-                style={{ color: '#9ca3af', fontSize: '0.9rem', flexShrink: 0 }}
-              />
-              <div style={{ display: 'flex', flexDirection: 'column', flex: 1 }}>
-                <span className={styles.heroPeriodoCampoLabel}>Desde</span>
-                <input
-                  ref={desdeRef}
-                  id="desde"
-                  type="date"
-                  className={styles.heroPeriodoInput}
-                  value={desde}
-                  onChange={(e) => setDesde(e.target.value)}
-                />
-              </div>
-            </div>
-            <span style={{ color: '#d1d5db', flexShrink: 0 }}>—</span>
-            <div className={styles.heroPeriodoCampo} onClick={() => hastaRef.current?.showPicker()}>
-              <i
-                className="bi bi-calendar3"
-                style={{ color: '#9ca3af', fontSize: '0.9rem', flexShrink: 0 }}
-              />
-              <div style={{ display: 'flex', flexDirection: 'column', flex: 1 }}>
-                <span className={styles.heroPeriodoCampoLabel}>Hasta</span>
-                <input
-                  ref={hastaRef}
-                  id="hasta"
-                  type="date"
-                  className={styles.heroPeriodoInput}
-                  value={hasta}
-                  onChange={(e) => setHasta(e.target.value)}
-                />
-              </div>
-            </div>
-            <button
-              className={styles.aplicarBtn}
-              onClick={aplicar}
-              disabled={!!rangoInvalido || recargando}
-            >
-              {recargando ? (
-                <span className="spinner-border spinner-border-sm" role="status" />
-              ) : (
-                'Aplicar'
-              )}
-            </button>
-          </div>
-          {rangoInvalido && (
-            <span style={{ color: '#fca5a5', fontSize: '0.82rem' }}>
-              &quot;Desde&quot; no puede ser posterior a &quot;Hasta&quot;.
-            </span>
-          )}
+          <FiltroPeriodo
+            desde={desde}
+            setDesde={setDesde}
+            hasta={hasta}
+            setHasta={setHasta}
+            rangoInvalido={rangoInvalido}
+            aplicar={aplicar}
+            recargando={recargando}
+          />
         </div>
       </div>
 
@@ -155,65 +102,10 @@ const Administrador = () => {
 
             <div className="row g-3 mb-3">
               <div className="col-12 col-md-6">
-                <div className={styles.propuestasCard}>
-                  <div className={styles.propuestasBarras}>
-                    <p className={styles.seccionTitulo}>
-                      <i
-                        className="bi bi-pie-chart-fill"
-                        style={{ color: 'var(--color-secondary)' }}
-                      />
-                      Propuestas por estado
-                    </p>
-                    <div className="d-flex flex-column gap-3">
-                      {PROPUESTAS.map(({ label, key, color }) => {
-                        const valor = stats.propuestasPorEstado?.[key] ?? 0
-                        const total = stats.totalPropuestas
-                        const pct = total > 0 ? Math.min(100, Math.round((valor / total) * 100)) : 0
-                        return (
-                          <div key={key} className="d-flex flex-column gap-1">
-                            <div className="d-flex justify-content-between align-items-center">
-                              <div className="d-flex align-items-center gap-2">
-                                <span
-                                  style={{
-                                    width: 9,
-                                    height: 9,
-                                    borderRadius: '50%',
-                                    background: color,
-                                    display: 'inline-block',
-                                    flexShrink: 0,
-                                  }}
-                                />
-                                <span style={{ fontSize: '0.85rem', color: 'var(--color-text)' }}>
-                                  {label}
-                                </span>
-                              </div>
-                              <span style={{ fontSize: '0.85rem', fontWeight: 600 }}>{valor}</span>
-                            </div>
-                            <div style={{ background: '#e9ecef', borderRadius: 4, height: 7 }}>
-                              <div
-                                style={{
-                                  width: `${pct}%`,
-                                  background: color,
-                                  borderRadius: 4,
-                                  height: '100%',
-                                  transition: 'width 0.4s ease',
-                                }}
-                              />
-                            </div>
-                          </div>
-                        )
-                      })}
-                    </div>
-                  </div>
-                  <div className={styles.totalPropuestasCard}>
-                    <i
-                      className="bi bi-tags-fill"
-                      style={{ color: '#d49a2c', fontSize: '1.4rem' }}
-                    />
-                    <div className={styles.totalPropuestasNumero}>{stats.totalPropuestas}</div>
-                    <div className={styles.totalPropuestasLabel}>Propuestas realizadas</div>
-                  </div>
-                </div>
+                <PropuestasPorEstado
+                  propuestasPorEstado={stats.propuestasPorEstado}
+                  totalPropuestas={stats.totalPropuestas}
+                />
               </div>
 
               <div className="col-12 col-md-6">
